Move merge_segments to BaseVad, fix docstrings, update VAD tests
