+++ src/onnx_asr/models/silero.py
@@ -17,8 +17,6 @@ from onnx_asr.vad import BaseVad
 class SileroVad(BaseVad):
     """Silero VAD implementation."""
 
-    INF = 10**15
-
     def __init__(self, model_files: dict[str, Path], onnx_options: OnnxSessionOptions):
         """Create Silero VAD.
 
@@ -86,35 +84,6 @@ class SileroVad(BaseVad):
             elif state == 1 and p < neg_threshold:
                 state = 0
                 yield start, i * hop_size
-
-    def _merge_segments(
-        self,
-        segments: Iterator[tuple[int, int]],
-        waveform_len: int,
-        sample_rate: int,
-        *,
-        min_speech_duration_ms: float = 250,
-        max_speech_duration_s: float = 20,
-        min_silence_duration_ms: float = 100,
-        speech_pad_ms: float = 30,
-        **kwargs: float,
-    ) -> Iterator[tuple[int, int]]:
-        speech_pad = int(speech_pad_ms * sample_rate // 1000)
-        min_speech_duration = int(min_speech_duration_ms * sample_rate // 1000) - 2 * speech_pad
-        max_speech_duration = int(max_speech_duration_s * sample_rate) - 2 * speech_pad
-        min_silence_duration = int(min_silence_duration_ms * sample_rate // 1000) + 2 * speech_pad
-
-        cur_start, cur_end = -self.INF, -self.INF
-        for start, end in chain(segments, ((waveform_len, waveform_len), (self.INF, self.INF))):
-            if start - cur_end < min_silence_duration and end - cur_start < max_speech_duration:
-                cur_end = end
-            else:
-                if cur_end - cur_start > min_speech_duration:
-                    yield max(cur_start - speech_pad, 0), min(cur_end + speech_pad, waveform_len)
-                while end - start > max_speech_duration:
-                    yield max(start - speech_pad, 0), start + max_speech_duration + speech_pad
-                    start += max_speech_duration
-                cur_start, cur_end = start, end
 
     def segment_batch(
         self,
