+++ src/onnx_asr/vad.py
@@ -3,7 +3,7 @@
 from abc import abstractmethod
 from collections.abc import Iterator
 from dataclasses import dataclass
-from itertools import islice
+from itertools import chain, islice
 from typing import Literal, Protocol
 
 import numpy as np
@@ -49,6 +49,37 @@ class Vad(Protocol):
 
 class BaseVad(Vad):
     """Base VAD class."""
+
+    INF = 10**15
+
+    def _merge_segments(
+        self,
+        segments: Iterator[tuple[int, int]],
+        waveform_len: int,
+        sample_rate: int,
+        *,
+        min_speech_duration_ms: float = 250,
+        max_speech_duration_s: float = 20,
+        min_silence_duration_ms: float = 100,
+        speech_pad_ms: float = 30,
+        **kwargs: float,
+    ) -> Iterator[tuple[int, int]]:
+        speech_pad = int(speech_pad_ms * sample_rate // 1000)
+        min_speech_duration = int(min_speech_duration_ms * sample_rate // 1000) - 2 * speech_pad
+        max_speech_duration = int(max_speech_duration_s * sample_rate) - 2 * speech_pad
+        min_silence_duration = int(min_silence_duration_ms * sample_rate // 1000) + 2 * speech_pad
+
+        cur_start, cur_end = -self.INF, -self.INF
+        for start, end in chain(segments, ((waveform_len, waveform_len), (self.INF, self.INF))):
+            if start - cur_end < min_silence_duration and end - cur_start < max_speech_duration:
+                cur_end = end
+            else:
+                if cur_end - cur_start > min_speech_duration:
+                    yield max(cur_start - speech_pad, 0), min(cur_end + speech_pad, waveform_len)
+                while end - start > max_speech_duration:
+                    yield max(start - speech_pad, 0), start + max_speech_duration + speech_pad
+                    start += max_speech_duration
+                cur_start, cur_end = start, end
 
     @abstractmethod
     def segment_batch(
